findr: passing ALL tests
This final step took a little work. The trick here was to ensure that
we tested for the possibility of the an empty vector for both
entry_types and names.

If names is empty, just treat it as a valid match.

Note: I was originally using `to_str()' followed by `unwrap_or("")',
but the book's solution was a lot cleaner using `to_string_lossy()'.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>